HIL

Module protocol

Protocol Module

This module handles the mapping between CSV simulation data and the VectorNav binary serial protocol. This module also handles the data output of the stm into the dedicated CSVs

Structs

AirbakesControlerOutput

This struct shall save the output of the airbrakes controler that runs on the stm. The serialize trait is aplied because we want to save the data to a CSV later

HorizontalFilterOutput

This struct shall save the output of the horizontal kalman filter that runs on the stm. The serialize trait is aplied because we want to save the data to a CSV later

VerticalFilterOutput

This struct shall save the output of the vertical kalman filter that runs on the stm.

VnSimulData

deserialize to directly map from csv to struct : fields as written in the CSV strut read from the csv and to be sent via serial port (excluding the headers and CRCs)

Constants

PAYLOAD_OFFSET 🔒

Offset calculation: 1 (Sync) + 1 (ID) + SAFE_BITS_LEN

Functions

calculate_vn_crc

Rust implementation of the VectorNav Datasheet CRC

is_packet_valid

Checks if calculated CRC is equal to received CRC

parse_controller

parse_horizontal

parse_vertical