HIL::protocol

Struct VnSimulData

pub struct VnSimulData {

    pub Yaw: f32,
    pub Pitch: f32,
    pub Roll: f32,
    pub UncompAccX: f32,
    pub UncompAccY: f32,
    pub UncompAccZ: f32,
    pub UncompGyroX: f32,
    pub UncompGyroY: f32,
    pub UncompGyroZ: f32,
    pub Temperature: f32,
    pub Pressure: f32,
    pub MagX: f32,
    pub MagY: f32,
    pub MagZ: f32,
    pub AccelX: f32,
    pub AccelY: f32,
    pub AccelZ: f32,
    pub GyroX: f32,
    pub GyroY: f32,
    pub GyroZ: f32,
    pub QuatX: f32,
    pub QuatY: f32,
    pub QuatZ: f32,
    pub QuatS: f32,
    pub LinAccelX: f32,
    pub LinAccelY: f32,
    pub LinAccelZ: f32,
    pub LinAccelN: f32,
    pub LinAccelE: f32,
    pub LinAccelD: f32,

}

deserialize to directly map from csv to struct : fields as written in the CSV strut read from the csv and to be sent via serial port (excluding the headers and CRCs)

c char array : [8] to [127]

Fields
Yaw: f32Pitch: f32Roll: f32UncompAccX: f32UncompAccY: f32UncompAccZ: f32UncompGyroX: f32UncompGyroY: f32UncompGyroZ: f32Temperature: f32Pressure: f32MagX: f32MagY: f32MagZ: f32AccelX: f32AccelY: f32AccelZ: f32GyroX: f32GyroY: f32GyroZ: f32QuatX: f32QuatY: f32QuatZ: f32QuatS: f32LinAccelX: f32LinAccelY: f32LinAccelZ: f32LinAccelN: f32LinAccelE: f32LinAccelD: f32
Implementations

impl VnSimulData

we use this to convert the data from the structure into a X-byte VectorNav binary packet

pub fn packet_to_vn(&self) -> Vec<u8>

Converts the structure data into a 130-byte VectorNav binary packet.

The first 8 bytes are headers (Sync, Group selection, etc.). See the VN Control Center to verify these header bytes.

Return

Returns a Vec<u8> of length 130 ready for serial transmission.