HIL

Module serial_utils

Serial utlis Module

This module handles opening the serial port (ttyACM, we are using the ST-Link as the bridge between the PC and the STM)

Functions

get_serial_port

Open serial port In Rust, the SerialPort object has a Drop implementation. When main finishes, Rust identifies that port is no longer needed and runs its internal “cleanup” code automatically, which releases the system handle for /dev/ttyUSB